componentized password output field
