+++ src/components/form/PasswordOutput.vue
@@ -1,0 +1,60 @@
+<script setup lang="ts">
+  import { ref, type Ref } from 'vue'
+  import { type Composer, useI18n } from 'vue-i18n'
+
+  type Props = {
+    password: string
+  }
+
+  defineProps<Props>()
+
+  const { t }: Composer = useI18n({ useScope: 'global' })
+  const copyToClipboardError: Ref<string> = ref('')
+  const passwordCopied: Ref<boolean> = ref(false)
+  const showPassword: Ref<boolean> = ref(false)
+
+  function copyToClipboard(text: string): void {
+    navigator.clipboard.writeText(text).then(
+      () => {
+        passwordCopied.value = true
+        copyToClipboardError.value = ''
+      },
+      (error: unknown) => {
+        // eslint-disable-next-line no-console
+        console.log(error)
+        copyToClipboardError.value = t('admin.person.copyPasswordError')
+      }
+    )
+  }
+</script>
+
+<template>
+  <v-text-field
+    v-if="password"
+    data-testid="password-output-field"
+    :error-messages="copyToClipboardError"
+    :hint="passwordCopied ? $t('admin.person.copyPasswordSuccess') : ''"
+    :persistent-hint="passwordCopied"
+    readonly
+    :type="showPassword ? 'text' : 'password'"
+    :value="password"
+    variant="outlined"
+  >
+    <template v-slot:append-inner>
+      <v-icon
+        @click.stop="showPassword = !showPassword"
+        data-testid="show-password-icon"
+        :icon="showPassword ? 'mdi-eye' : 'mdi-eye-off'"
+      ></v-icon>
+    </template>
+    <template v-slot:append>
+      <v-icon
+        @click.stop="copyToClipboard(password)"
+        data-testid="copy-password-icon"
+        icon="mdi-content-copy"
+      ></v-icon>
+    </template>
+  </v-text-field>
+</template>
+
+<style></style>
